Make a bunch of things public so that miri can use them
